refact: #55 add @common-components path alias & move src/components/common to src/components/_common
